requirements.txt
flask
pandas
numpy
scikit-learn
joblib
PyPDF2
pillow
pytesseract